refactor: extract form logic to standalone script and fix tests
