+++ src/components/Form.jsx
@@ -1,132 +1,25 @@
 /** @jsxRuntime automatic */
 /** @jsxImportSource hono/jsx */
 import { CustomRules } from './CustomRules.jsx';
+import { formScriptFn } from './formScript.js';
 import { UNIFIED_RULES, PREDEFINED_RULE_SETS } from '../config.js';
 
 export const Form = (props) => {
   const { t } = props;
 
-  const scriptContent = `
-        window.formData = function() {
-          return {
-            input: '',
-            showAdvanced: false,
-            selectedRules: [],
-            selectedPredefinedRule: 'balanced',
-            groupByCountry: false,
-            enableClashUI: false,
-            externalController: '',
-            externalUiDownloadUrl: '',
-            configType: 'singbox',
-            configEditor: '',
-            customUA: '',
-            loading: false,
-            generatedLinks: null,
-            processingText: ${JSON.stringify(t('processing'))},
-            convertText: ${JSON.stringify(t('convert'))},
-            
-            init() {
-              // Load saved data
-              this.input = localStorage.getItem('inputTextarea') || '';
-              this.showAdvanced = localStorage.getItem('advancedToggle') === 'true';
-              this.groupByCountry = localStorage.getItem('groupByCountry') === 'true';
-              this.enableClashUI = localStorage.getItem('enableClashUI') === 'true';
-              this.externalController = localStorage.getItem('externalController') || '';
-              this.externalUiDownloadUrl = localStorage.getItem('externalUiDownloadUrl') || '';
-              this.customUA = localStorage.getItem('userAgent') || '';
-              this.configEditor = localStorage.getItem('configEditor') || '';
-              this.configType = localStorage.getItem('configType') || 'singbox';
-              
-              // Initialize rules
-              this.applyPredefinedRule();
-              
-              // Watchers to save state
-              this.$watch('input', val => localStorage.setItem('inputTextarea', val));
-              this.$watch('showAdvanced', val => localStorage.setItem('advancedToggle', val));
-              this.$watch('groupByCountry', val => localStorage.setItem('groupByCountry', val));
-              this.$watch('enableClashUI', val => localStorage.setItem('enableClashUI', val));
-              this.$watch('externalController', val => localStorage.setItem('externalController', val));
-              this.$watch('externalUiDownloadUrl', val => localStorage.setItem('externalUiDownloadUrl', val));
-              this.$watch('customUA', val => localStorage.setItem('userAgent', val));
-              this.$watch('configEditor', val => localStorage.setItem('configEditor', val));
-              this.$watch('configType', val => localStorage.setItem('configType', val));
-            },
-            
-            applyPredefinedRule() {
-              if (this.selectedPredefinedRule === 'custom') return;
-              
-              const rules = ${JSON.stringify(PREDEFINED_RULE_SETS)};
-              if (rules && rules[this.selectedPredefinedRule]) {
-                this.selectedRules = rules[this.selectedPredefinedRule];
-              }
-            },
-            
-            saveBaseConfig() {
-              localStorage.setItem('configEditor', this.configEditor);
-              localStorage.setItem('configType', this.configType);
-              alert(${JSON.stringify(t('saveConfigSuccess'))});
-            },
-            
-            clearBaseConfig() {
-              if (confirm(${JSON.stringify(t('confirmClearConfig'))})) {
-                this.configEditor = '';
-                localStorage.removeItem('configEditor');
-              }
-            },
-            
-            clearAll() {
-              if (confirm(${JSON.stringify(t('confirmClearAll'))})) {
-                this.input = '';
-                this.generatedLinks = null;
-              }
-            },
-            
-            async submitForm() {
-              this.loading = true;
-              try {
-                // Get custom rules from the child component via the hidden input
-                const customRulesInput = document.querySelector('input[name="customRules"]');
-                const customRules = customRulesInput && customRulesInput.value ? JSON.parse(customRulesInput.value) : [];
-                
-                // Construct URLs
-                const origin = window.location.origin;
-                const params = new URLSearchParams();
-                params.append('config', this.input);
-                params.append('ua', this.customUA);
-                params.append('selectedRules', JSON.stringify(this.selectedRules));
-                params.append('customRules', JSON.stringify(customRules));
-                
-                if (this.groupByCountry) params.append('group_by_country', 'true');
-                if (this.enableClashUI) params.append('enable_clash_ui', 'true');
-                if (this.externalController) params.append('external_controller', this.externalController);
-                if (this.externalUiDownloadUrl) params.append('external_ui_download_url', this.externalUiDownloadUrl);
-                
-                const queryString = params.toString();
+  const translations = {
+    processing: t('processing'),
+    convert: t('convert'),
+    saveConfigSuccess: t('saveConfigSuccess'),
+    confirmClearConfig: t('confirmClearConfig'),
+    confirmClearAll: t('confirmClearAll'),
+    errorGeneratingLinks: t('errorGeneratingLinks')
+  };
 
-                this.generatedLinks = {
-                  xray: origin + '/xray?' + queryString,
-                  singbox: origin + '/singbox?' + queryString,
-                  clash: origin + '/clash?' + queryString,
-                  surge: origin + '/surge?' + queryString
-                };
-                
-                // Scroll to results
-                setTimeout(() => {
-                  const resultsDiv = document.querySelector('.mt-12');
-                  if (resultsDiv) {
-                    resultsDiv.scrollIntoView({ behavior: 'smooth' });
-                  }
-                }, 100);
-                
-              } catch (error) {
-                console.error('Error generating links:', error);
-                alert(${JSON.stringify(t('errorGeneratingLinks'))});
-              } finally {
-                this.loading = false;
-              }
-            }
-          }
-        }
+  const scriptContent = `
+    window.APP_TRANSLATIONS = ${JSON.stringify(translations)};
+    window.PREDEFINED_RULE_SETS = ${JSON.stringify(PREDEFINED_RULE_SETS)};
+    (${formScriptFn.toString()})();
   `;
 
   return (
